Relocate DictionaryComparer to SIL.XForge
