+++ src/SIL.XForge/Utils/DictionaryComparer.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace SIL.XForge.Scripture.Services;
+namespace SIL.XForge.Utils;
 
 public class DictionaryComparer<TKey, TValue> : IEqualityComparer<Dictionary<TKey, TValue>>
 {
